src/backend/todo_service/requirements.txt
pydantic
fastapi
uvicorn
mysql-connector-python